requirements.txt
streamlit>=1.28.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
PyPDF2>=3.0.0
lxml>=4.9.0

